refactor: move MyContracts local state into Redux store.myContracts.contracts
