+++ package.json
@@ -74,6 +74,7 @@
     "uuid": "^8.3.2",
     "web3": "^1.6.1",
     "web3-core": "^1.7.0",
+    "web3-eth": "^1.7.1",
     "web3-eth-contract": "^1.7.1",
     "web3-utils": "^1.7.0",
     "yup": "^0.32.11"
