+++ src/store/myContracts/selectors.ts
@@ -1,0 +1,6 @@
+import { IContractsCard } from 'pages/MyContracts/MyContracts.helpers';
+import type { State } from 'types';
+
+export default {
+  getMyContracts: (state: State): IContractsCard[] => state.myContracts.contracts,
+};
