+++ src/store/rootReducer.ts
@@ -1,9 +1,11 @@
 import ui from './ui/reducer';
 import user from './user/reducer';
 import contractForms from './contractForms/reducer';
+import myContracts from './myContracts/reducer';
 
 export default {
   ui,
   user,
   contractForms,
+  myContracts,
 };
