+++ src/types/store/myContracts.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line import/no-cycle
+import { IContractsCard } from '../../pages/MyContracts/MyContracts.helpers';
+
 export type MyContractsState = {
-  contracts: [];
+  contracts: IContractsCard[];
 };
