+++ src/types/store/store.ts
@@ -1,6 +1,7 @@
 import type { Dispatch as DispatchReact } from 'react';
 import { ContractFormsState } from './contractForms';
 import { UserState } from './user';
+import { MyContractsState } from './myContracts';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Action<T, P = any, M = void> = { type: T; payload?: P; meta?: M };
@@ -9,4 +10,5 @@ export type Dispatch = DispatchReact<{ type: string }>;
 export type State = {
   user: UserState,
   contractForms: ContractFormsState,
+  myContracts: MyContractsState,
 };
