+++ src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './getDeepValueByPath';
 export * from './contractsHelper';
 export * from './dateTimeConverters';
 export * from './validationSchema.helpers';
+export * from './web3Objects.helpers';
